refactor decorator script (#31741)
